chore: move conditional AtomicU64 impl to new file
Keeping the implementation out of a macro lets rustfmt apply to it.
